+++ PlanIt/.gitignore
@@ -11,6 +11,7 @@ node_modules
 dist
 dist-ssr
 *.local
+*.iml
 
 # Editor directories and files
 .vscode/*
