+++ PlanIt/src/App.jsx
@@ -1,45 +1,33 @@
-import Register from "./pages/register/Regitser.jsx";
-import Login from "./pages/auth/Login.jsx";
-import AccountRecovery from "./pages/auth/accountRecovery.jsx";
-import PasswordRecovery from "./pages/auth/PasswordRecovery.jsx";
-import PostRoutes from "./routes/post.jsx";
 import {
-    BrowserRouter as Router,
-    Routes,
-    Route,
-    Navigate,
+  BrowserRouter as Router,
+  Route,
+  Routes,
+  Navigate,
 } from "react-router-dom";
-import WelcomeScreen from "./pages/register/WelcomeScreen.jsx";
-import SplashPage from "./pages/splash/SplashPage.jsx";
-import Inquiry from "./pages/splash/inquiry/Inquiry.jsx";
-import Findpwd from "./pages/splash/findpwd/Findpwd.jsx";
-import Community from "./pages/community/Community.jsx";
 import { Provider } from "react-redux";
 import { store } from "./pages/community/storage/commuStore/store.js";
 
+import Community from "./pages/community/Community.jsx";
+import AuthRoutes from "./routes/AuthRoutes.jsx";
+import InquiryRoutes from "./routes/InquiryRoutes.jsx";
+import PostRoutes from "./routes/post.jsx";
+
 function App() {
-    return (
-        <Provider store={store}>
-            <Router>
-                <Routes>
-                    <Route path="/" element={<Navigate to="/community/post" />} />
-                    <Route path="/community/*" element={<Community />} />
+  return (
+    <Provider store={store}>
+      <Router>
+        <Routes>
+          <Route path="/" element={<Navigate to="/community/post" />} />
+          <Route path="/community/*" element={<Community />} />
+          {/* <Route path="/" element={<Navigate to="/community/post" />} /> */}
 
-                    {/*<Route path="/" element={<Navigate to="/register/registerBody" />} />*/}
-                    {/*<Route path="/register/*" element={<Register />} />*/}
-
-                    {/* <Route path="/" element={<SplashPage/>}/> */}
-                    <Route path="/login" element={<Login />} />
-                    <Route path="/findAccount" element={<AccountRecovery />} />
-                    <Route path="/findPW" element={<PasswordRecovery />} />
-                    <Route path="/inquiry" element={<Inquiry />} />
-                    <Route path="/password" element={<Findpwd />} />
-                    <Route path="/welcome" element={<WelcomeScreen />} />
-                    {/* <Route path="/" element={<Register />} /> */}
-                </Routes>
-            </Router>
-        </Provider>
-    );
+          <Route path="/*" element={<AuthRoutes />} />
+          <Route path="/*" element={<InquiryRoutes />} />
+          <Route path="/*" element={<PostRoutes />} />
+        </Routes>
+      </Router>
+    </Provider>
+  );
 }
 
 export default App;
