+++ PlanIt/src/styles/variables.css
@@ -1,24 +1,27 @@
-@import url('https://cdn.jsdelivr.net/gh/orioncactus/pretendard/dist/web/static/pretendard.css');
+@import url("https://cdn.jsdelivr.net/gh/orioncactus/pretendard@v1.3.9/dist/web/static/pretendard.min.css");
 
 :root {
-    --primary-color: #2988FF;
+  --primary-color: #2988ff;
 
-    /* border */
-    --border-default: #D9D9D9;
-    --border-neutral-secondary: #B3B3B3;
+  /* border */
+  --border-default: #d9d9d9;
+  --border-neutral-secondary: #b3b3b3;
 
-    /* background */
-    --background-default-default-tap: #F5F5F5;
-    --background-default-secondary: #F5F5F5;
+  /* background */
+  --background-default-default-tap: #f5f5f5;
+  --background-default-secondary: #f5f5f5;
 
-    /* 폰트 */
-    --font-family-body: 'Pretendard', -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, 'Helvetica Neue', Arial, 'Noto Sans', sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji', 'Segoe UI Symbol', 'Noto Color Emoji';
-    --body-caption-12px: 12px;
-    --body-secondary-16px: 16px;
-    --text-default: #1E1E1E;
-    --text-default-secondary: #808080;
+  /* 폰트 */
+  --font-family-body:
+    "Pretendard", -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto,
+    "Helvetica Neue", Arial, "Noto Sans", sans-serif, "Apple Color Emoji",
+    "Segoe UI Emoji", "Segoe UI Symbol", "Noto Color Emoji";
+  --body-caption-12px: 12px;
+  --body-secondary-16px: 16px;
+  --text-default: #1e1e1e;
+  --text-default-secondary: #808080;
 }
 
 .error {
-    color: red;
+  color: red;
 }
